rag_service/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
pydantic
python-multipart
chromadb
python-dotenv
pypdf
pandas
openpyxl
python-docx
redis
supabase

